refactor(getValidatedCarNames): 자동차 이름 입력과 유효성 검사를 getValidatedCarNames 함수로 분리
run 메서드가 전체 흐름만 담당하도록 리팩토링
