+++ src/App.js
@@ -1,29 +1,22 @@
-import { splitByComma, joinByComma } from './utils/stringUtils.js';
+import { joinByComma } from './utils/stringUtils.js';
 import Car from './car/Car.js';
-import Game from './racingGame/RacingGame.js';
+import RacingGame from './racingGame/RacingGame.js';
 import { getUserInput, printKeyValueFormat } from './utils/ioModule.js';
-import {
-  validateCarNames,
-  validateRacingCarInput,
-  validateTryCount,
-} from './validator/validatePipeline.js';
+import { validateTryCount } from './validator/validatePipeline.js';
 import { PRINT_MESSAGES } from './constants/messages.js';
+import getValidatedCarNames from './functions/getValidatedCarNames.js';
 
 class App {
   async run() {
-    const inputForRacingCars = await getUserInput(
-      PRINT_MESSAGES.INPUT.CAR_NAME,
-    );
-    validateRacingCarInput(inputForRacingCars);
-    const carNamesArray = splitByComma(inputForRacingCars);
-    validateCarNames(carNamesArray);
+    const carNamesArray = await getValidatedCarNames();
+
     const inputForTryCount = await getUserInput(PRINT_MESSAGES.INPUT.TRY_COUNT);
     validateTryCount(inputForTryCount);
 
     const racingCars = carNamesArray.map((name) => new Car(name));
-    const game = new Game(racingCars, Number(inputForTryCount));
-    game.play();
-    const winners = game.getWinnerNames();
+    const racingGame = new RacingGame(racingCars, Number(inputForTryCount));
+    racingGame.play();
+    const winners = racingGame.getWinnerNames();
     printKeyValueFormat(PRINT_MESSAGES.OUTPUT.WINNER, joinByComma(winners));
   }
 }
